Cập nhật module mới

+++ app/Http/Controllers/Backend/V2/Major/MajorController.php
@@ -5,9 +5,8 @@ use Illuminate\Http\Request;
 use App\Http\Requests\Major\Major\StoreRequest;
 use App\Http\Requests\Major\Major\UpdateRequest;
 use App\Services\V2\Impl\Major\MajorService;
+use App\Services\V2\Impl\Scholar\TrainService;
 use App\Models\Language;
-use Illuminate\Pagination\LengthAwarePaginator;
-use App\Http\Resources\MajorResource;
 use App\Services\V2\Impl\Major\MajorCatalogueService;
 
 class MajorController extends Controller {
@@ -15,15 +14,18 @@ class MajorController extends Controller {
 
     private $service;
     private $majorCatalogueService;
+    private $trainService;
     protected $language;
 
     public function __construct(
         MajorService $service,
-        MajorCatalogueService $majorCatalogueService
+        MajorCatalogueService $majorCatalogueService,
+        TrainService $trainService
     )
     {
         $this->service = $service;
         $this->majorCatalogueService = $majorCatalogueService;
+        $this->trainService = $trainService;
         $this->middleware(function($request, $next){
             $locale = app()->getLocale();
             $language = Language::where('canonical', $locale)->first();
@@ -33,7 +35,7 @@ class MajorController extends Controller {
     }
 
     public function index(Request $request){
-         // $this->authorize('modules', 'major.major.index');
+        $this->authorize('modules', 'major.index');
         $majors = $this->service->pagination($request);
         $config = [
             'model' => 'Major',
@@ -49,7 +51,7 @@ class MajorController extends Controller {
     }
 
     public function create(){
-         // $this->authorize('modules', 'major.major.create');
+        $this->authorize('modules', 'major.create');
         $config = [
             'model' => 'Major',
             'seo' => $this->seo(),
@@ -57,16 +59,18 @@ class MajorController extends Controller {
             'extendJs' => true
         ];
         $dropdown = $this->majorCatalogueService->dropdown();
+        $trains = $this->trainService->all()->pluck('name', 'id');
         $template = 'backend.major.major.store';
         return view('backend.dashboard.layout', compact(
             'template',
             'config',
-            'dropdown'
+            'dropdown',
+            'trains'
         ));
     }
 
     public function edit($id){
-         // $this->authorize('modules', 'major.major.update');
+        $this->authorize('modules', 'major.update');
         if(!$major = $this->service->findById($id)){
             return redirect()->route('major.major.index')->with('error','Bản ghi không tồn tại'); 
         }
@@ -77,12 +81,14 @@ class MajorController extends Controller {
             'extendJs' => true
         ];
         $dropdown = $this->majorCatalogueService->dropdown();
+        $trains = $this->trainService->all()->pluck('name', 'id');
         $template = 'backend.major.major.store';
         return view('backend.dashboard.layout', compact(
             'template',
             'config',
             'major',
-            'dropdown'
+            'dropdown',
+            'trains'
         ));     
     }
     
@@ -94,7 +100,7 @@ class MajorController extends Controller {
     }
 
     public function update($id, UpdateRequest $request){
-         // $this->authorize('modules', 'major.major.update');
+         $this->authorize('modules', 'major.update');
         if($response = $this->service->save($request, 'update', $id)){
             return redirect()->back()->with('success', 'Cập nhật bản ghi thành công');
         }
@@ -102,7 +108,7 @@ class MajorController extends Controller {
     }
 
     public function delete($id){
-        //  $this->authorize('modules', 'major.major.destroy');
+         $this->authorize('modules', 'major.destroy');
         if(!$major = $this->service->findById($id)){
             return redirect()->route('major.major.index')->with('error','Bản ghi không tồn tại'); 
         }
@@ -120,7 +126,7 @@ class MajorController extends Controller {
     }
 
     public function destroy($id){
-        //  $this->authorize('modules', 'major.major.destroy');
+         $this->authorize('modules', 'major.destroy');
         if($response = $this->service->destroy($id)){
             return redirect()->route('major.major.index')->with('success', 'Xóa bản ghi thành công');
         }
@@ -131,17 +137,17 @@ class MajorController extends Controller {
     private function seo(){
         return [
             'index' => [
-                'title' => 'Quản lý Chuyên Ngành',
-                'table' => 'Danh sách Chuyên Ngành'
+                'title' => 'Quản lý chuyên Ngành',
+                'table' => 'Danh sách chuyên Ngành'
             ],
             'create' => [
-                'title' => 'Thêm mới Chuyên Ngành'
+                'title' => 'Thêm mới chuyên Ngành'
             ],
             'update' => [
-                'title' => 'Cập nhật Chuyên Ngành'
+                'title' => 'Cập nhật chuyên Ngành'
             ],
             'delete' => [
-                'title' => 'Xóa Chuyên Ngành'
+                'title' => 'Xóa chuyên Ngành'
             ]
         ];
     }
